RobotFramework/requirements.txt
requests==2.24.0
robotframework==3.2.1
robotframework-appiumlibrary==1.5.0.6
robotframework-databaselibrary==1.2.4
robotframework-debuglibrary==2.2.1
robotframework-requests==0.7.0
robotframework-sshlibrary==3.4.0